refactor: move contracts to types.ts
Consolidate all TypeScript interfaces and types from /src/contracts directory
into a single /src/types.ts file and remove the contracts folder.

Changes:
- Created src/types.ts with merged content from contracts/index.ts and contracts/rule.ts
- Updated all imports from '../contracts' to '../types'
- Added export * from './types' to main index.ts
- Deleted src/contracts/ directory

Benefits:
- Simpler file structure (single file vs directory)
- Follows TypeScript conventions (types.ts pattern)
- Easier type discovery and maintenance
- All types consolidated in one location
- No breaking changes to public API

Closes #16
