+++ src/types.ts
@@ -1,4 +1,63 @@
-import { Rule } from './rule';
+/**
+ * List of Trivule rules grouped by Rule type
+ */
+export type Rule =
+  | 'required'
+  | 'email'
+  | 'maxlength'
+  | 'minlength'
+  | 'min'
+  | 'max'
+  | 'string'
+  | 'between'
+  | 'startWith'
+  | 'endWith'
+  | 'contains'
+  | 'in'
+  | 'integer'
+  | 'int'
+  | 'number'
+  | 'len'
+  | 'length'
+  | 'numeric'
+  | 'maxFileSize'
+  | 'minFileSize'
+  | 'size'
+  | 'file'
+  | 'boolean'
+  | 'startWithUpper'
+  | 'nullable'
+  | 'startWithLower'
+  | 'password'
+  | 'date'
+  | 'before'
+  | 'after'
+  | 'phone'
+  | 'url'
+  | 'time'
+  | 'startWithString'
+  | 'endWithString'
+  | 'hasLetter'
+  | 'excludes'
+  | 'regex'
+  | 'lower'
+  | 'upper'
+  | 'stringBetween'
+  | 'modulo'
+  | 'mod'
+  | 'only'
+  | 'mimes'
+  | 'digit'
+  | 'minDigit'
+  | 'maxDigit'
+  | 'lessThan'
+  | 'greaterThan'
+  | 'gthan'
+  | 'lthan'
+  | 'fileBetween'
+  | 'dateBetween'
+  | 'numberBetween';
+
 export type ITrivuleInputCallback<P> = (param: P) => void;
 
 /**
@@ -86,11 +145,9 @@ export type RulesBag = {
 export type RulesMessages = {
   [key: string]: string;
 };
-import { Trivule, TrivuleForm } from '../validation';
 
 /**
  * Represents a CSS selector that can be either an HTMLElement or a string.
- * @typedef {HTMLElement | string} CssSelector
  */
 export type CssSelector = HTMLElement | string;
 
@@ -242,7 +299,7 @@ export type TrivuleFormConfig = {
 
 export type TrivuleHooks = 'before.init' | 'after.init' | 'destroy' | string;
 
-export type TrivuleFormHandler = (tr: TrivuleForm) => unknown;
+export type TrivuleFormHandler = (tr: unknown) => unknown;
 
 /**
  * Service interface for extending Trivule functionality
@@ -252,13 +309,13 @@ export interface TrivuleService {
    * Register service with Trivule instance
    * @param trivule The Trivule instance
    */
-  register(trivule: typeof Trivule): void;
+  register(trivule: unknown): void;
 }
 
 /**
  * Closure-based service function type
  */
-export type TrivuleServiceFunction = (trivule: typeof Trivule) => void;
+export type TrivuleServiceFunction = (trivule: unknown) => void;
 
 export type TrivuleAttribute =
   | 'rules'
@@ -274,5 +331,3 @@ export type TrivuleAttribute =
   | 'name'
   | 'lang'
   | 'show';
-
-export * from './rule';
