+++ src/validation/tr-bag.ts
@@ -1,7 +1,7 @@
 import { dateBetween } from './../rules/date';
 import { endWithString, stringBetween } from '../rules/string';
 import { fileBetween, isMimes, minFileSize } from '../rules/file';
-import { Rule, RuleCallBack, RulesBag, RulesMessages } from '../contracts';
+import { Rule, RuleCallBack, RulesBag, RulesMessages } from '../types';
 import {
   between,
   contains,
