+++ src/validation/tr-input.ts
@@ -9,7 +9,7 @@ import {
   TrivuleHooks,
   TrivuleInputParms,
   ValidatableInput,
-} from '../contracts';
+} from '../types';
 import { AbstractInputralidator } from './abstract-input';
 import { TrBag } from './tr-bag';
 import { TrParameter } from './utils/parameter';
@@ -343,19 +343,6 @@ export class TrivuleInput extends AbstractInputralidator {
   removeNativeAttribute(attrName: string): this {
     this.inputElement?.removeAttribute(attrName);
     return this;
-  }
-  /**
-   * Retrieves the value of a native attribute from the Trivule input element.
-   *
-   * This method attempts to get the attribute value using the `getAttribute` method
-   * of the underlying `inputElement`. If the element doesn't have the attribute,
-   * it returns `undefined`.
-   *
-   * @param attrName The name of the attribute to retrieve.
-   * @returns The value of the attribute if it exists, otherwise `undefined`.
-   */
-  getNativeAttribute(attrName: string) {
-    return this.inputElement?.getAttribute(attrName);
   }
 
   /**
