+++ src/validation/trivule.ts
@@ -4,7 +4,7 @@ import {
   TrivuleFormConfig,
   TrivuleInputParms,
   ValidatableForm,
-} from '../contracts';
+} from '../types';
 import { TrConfig } from '../tr.config';
 import { TrBag } from './tr-bag';
 import { TrivuleForm } from './tr-form';
